支持CONNECT（https etc.）方法

+++ bin/stop.py
@@ -6,9 +6,10 @@ with open(sys.path[0]+'/pid','r') as f:
     pid_list = f.read().split('|')
 
 for pid in pid_list:
-    os.system('kill '+pid)
-
+    if pid:
+        os.system('kill '+pid)
+        print('stop process '+pid+' ok')
+    else:
+        print('nothing to stop')
 with open(sys.path[0]+'/pid','w') as f:
     f.truncate()
-    
-print('stop process '+'|'.join(pid_list)+' ok')
